refactor(llm): extract TextGenerationMixin — final facade shape (Tier-2.5 Task 4)
Move the ~17 text-generation methods (generate_response, generate_chat_response,
_build_completion_params, the hard-timeout/subprocess-isolation cluster,
observability, _make_request, prompt caching, multimodal image handling,
_is_temperature_restricted_model) plus _compute_cost_usd (retained method,
verbatim billing exception->None semantics — no separate cost module),
_MODEL_TIMEOUT_FLOOR_SECONDS, and the SUPPORTED_IMAGE_FORMATS /
TEMPERATURE_RESTRICTED_MODELS class constants into TextGenerationMixin
(_litellm_text_generation.py). Compose it first in the MRO:
LiteLLMClient(TextGenerationMixin, EmbeddingMixin, StructuredOutputMixin).

Per-mixin TYPE_CHECKING stubs (Tier-1b idiom) self-type the foreign members:
client-core config/logger/_api_key/_api_base/_api_version/_resolve_api_key (on
the facade) and the two cross-mixin edges resolved via MRO
(_provider_response_format + _maybe_parse_structured_output on
StructuredOutputMixin).

LLM-mock: every litellm call stays a shared-module attr (litellm.completion);
the subprocess worker is imported from _litellm_subprocess. SINK-1:
resolve_model_name is imported into this module for the model-role path (no test
patches the text-gen binding; the :717 embedding site was repointed in Task 2).

The facade is now the client-core shell: __init__ (keeps import litellm +
suppress_debug_info + the Braintrust callbacks init — SINK-3), _resolve_api_key,
_resolve_by_prefix (+ _SIMPLE_PROVIDER_PREFIXES), update_config, get_model,
get_config, create_litellm_client, __all__, and the identity re-exports of every
moved public + test-internal name. Bodies verbatim; all LLM suites +
billing/cost + mock-compliance snapshots green.
